Complete workshop 5: fix things in video
After the video I drew the following lessons:

* Instead of explicitly removing a vertical scrollbar that appeared
when a horizontal one was inserted the fixed height of the Header
could be removed, and thus stretch a little more, which also gave
the added benefit that a bigger font size could be used without
screwing up this part of the site.
* The items in the mobile menu did not quite position perfectly
as I'd though. To make the main navigation links appear exactly
in the middle a Filler was used, together with the footer, which
both grew with flex: 1. This squeezed the nav links to the middle.
I also changed the case when the menu is opened on a tablet: Instead
of setting the width to 80 % I now used the fixed width from figma,
which looks a lot better
* As per the solution video, I changed my breadcrumbs to be their
own component to make it more DRY and grouping the Mobilecrumbs
and the Title together instead of also duplicating the Title.
* Finally I also missed a minor detail that the filter-select was
floating too high above the sneakers in tablet mode. With a media
query together with align-items: center, everything looked a lot
better.